rename package local to shipper
