+++ pkg/storage/stores/shipper/boltdb_index_client.go
@@ -1,4 +1,4 @@
-package local
+package shipper
 
 import (
 	"context"
