+++ pkg/storage/stores/shipper/boltdb_shipper_table_client_test.go
@@ -1,4 +1,4 @@
-package local
+package shipper
 
 import (
 	"bytes"
