+++ pkg/storage/stores/shipper/filescollection.go
@@ -1,4 +1,4 @@
-package local
+package shipper
 
 import (
 	"context"
