+++ pkg/storage/stores/shipper/metrics.go
@@ -1,4 +1,4 @@
-package local
+package shipper
 
 import (
 	"sync"
