+++ pkg/storage/stores/shipper/shipper.go
@@ -1,4 +1,4 @@
-package local
+package shipper
 
 import (
 	"context"
